refactor(optimizer): centralize Pareto helpers for GEPA/GEPA-Flow
Extract shared multi-objective utilities to paretoUtils and replace inline duplicates in gepa.ts and gepaFlow.ts. No functional changes; simplifies maintenance and sets up for paper parity.
