+++ src/ax/dsp/optimizers/gepa.ts
@@ -16,6 +16,15 @@ import {
 } from '../optimizer.js';
 import type { AxGenOut } from '../types.js';
 import { ax } from '../template.js';
+import {
+  buildParetoFront,
+  hypervolume2D,
+  dominatesVectorEps,
+  average,
+  randomSubset,
+  selectCandidatePareto,
+  avgVec,
+} from './paretoUtils.js';
 
 /** Single-module GEPA (reflective prompt evolution with Pareto sampling) */
 export class AxGEPA extends AxBaseOptimizer {
@@ -104,24 +113,6 @@ export class AxGEPA extends AxBaseOptimizer {
         ? validationExamples
         : examples
     ).slice(0, this.paretoSetSize);
-
-    // Helper to average objective vectors
-    const avgVec = (
-      arrs: ReadonlyArray<Record<string, number>>
-    ): Record<string, number> => {
-      const sums: Record<string, number> = {};
-      const counts: Record<string, number> = {};
-      for (const r of arrs) {
-        for (const [k, v] of Object.entries(r)) {
-          sums[k] = (sums[k] || 0) + (typeof v === 'number' ? v : 0);
-          counts[k] = (counts[k] || 0) + 1;
-        }
-      }
-      const out: Record<string, number> = {};
-      for (const [k, s] of Object.entries(sums))
-        out[k] = s / Math.max(counts[k] || 1, 1);
-      return out;
-    };
 
     // Evaluate one example -> objective vector
     const evalOne = async (
@@ -601,221 +592,4 @@ export class AxGEPA extends AxBaseOptimizer {
       instructionA.length >= instructionB.length ? instructionA : instructionB
     ).slice(0, 2000);
   }
-}
-
-// --- Utilities ---
-
-function dominatesVector(
-  a: Readonly<Record<string, number>>,
-  b: Readonly<Record<string, number>>
-): boolean {
-  const keys = new Set([...Object.keys(a), ...Object.keys(b)]);
-  let atLeastAsGood = true;
-  let strictlyBetter = false;
-  for (const k of keys) {
-    const va = a[k] ?? 0;
-    const vb = b[k] ?? 0;
-    if (va < vb) {
-      atLeastAsGood = false;
-      break;
-    }
-    if (va > vb) strictlyBetter = true;
-  }
-  return atLeastAsGood && strictlyBetter;
-}
-
-function buildParetoFront(
-  items: ReadonlyArray<{
-    idx: number;
-    scores: Readonly<Record<string, number>>;
-  }>
-): Array<{
-  idx: number;
-  scores: Readonly<Record<string, number>>;
-  dominated: number;
-}> {
-  const front: Array<{
-    idx: number;
-    scores: Readonly<Record<string, number>>;
-    dominated: number;
-  }> = [];
-  for (let i = 0; i < items.length; i++) {
-    let dominatedCount = 0;
-    let isDominated = false;
-    for (let j = 0; j < items.length; j++) {
-      if (i === j) continue;
-      if (dominatesVector(items[j]!.scores, items[i]!.scores)) {
-        isDominated = true;
-        break;
-      }
-      if (dominatesVector(items[i]!.scores, items[j]!.scores)) dominatedCount++;
-    }
-    if (!isDominated)
-      front.push({
-        idx: items[i]!.idx,
-        scores: items[i]!.scores,
-        dominated: dominatedCount,
-      });
-  }
-  return front;
-}
-
-function computeCrowdingDistances(
-  front: ReadonlyArray<{
-    idx: number;
-    scores: Readonly<Record<string, number>>;
-  }>
-): Map<number, number> {
-  const dist = new Map<number, number>();
-  if (front.length === 0) return dist;
-  const keys = new Set<string>();
-  for (const f of front) for (const k of Object.keys(f.scores)) keys.add(k);
-  for (const f of front) dist.set(f.idx, 0);
-  for (const key of keys) {
-    const sorted = [...front].sort(
-      (a, b) => (a.scores[key] ?? 0) - (b.scores[key] ?? 0)
-    );
-    const min = sorted[0] ? (sorted[0].scores[key] ?? 0) : 0;
-    const max = sorted[sorted.length - 1]
-      ? (sorted[sorted.length - 1].scores[key] ?? 0)
-      : 0;
-    const range = Math.max(max - min, 1e-9);
-    if (sorted.length > 0) dist.set(sorted[0]!.idx, Number.POSITIVE_INFINITY);
-    if (sorted.length > 1)
-      dist.set(sorted[sorted.length - 1]!.idx, Number.POSITIVE_INFINITY);
-    for (let i = 1; i < sorted.length - 1; i++) {
-      const prev = sorted[i - 1]!.scores[key] ?? 0;
-      const next = sorted[i + 1]!.scores[key] ?? 0;
-      const inc = (next - prev) / range;
-      dist.set(sorted[i]!.idx, (dist.get(sorted[i]!.idx) ?? 0) + inc);
-    }
-  }
-  return dist;
-}
-
-function hypervolume2D(
-  front: ReadonlyArray<Readonly<Record<string, number>>>
-): number | undefined {
-  if (front.length === 0) return undefined;
-  // Detect objectives (use first vector)
-  const keys = Object.keys(front[0] ?? {});
-  if (keys.length !== 2) return undefined;
-  const [k1, k2] = keys;
-  // Sort by k1 descending
-  const sorted = [...front].sort((a, b) => (b[k1!] ?? 0) - (a[k1!] ?? 0));
-  let hv = 0;
-  let prevY = 0;
-  for (const p of sorted) {
-    const x = p[k1!] ?? 0;
-    const y = p[k2!] ?? 0;
-    const dy = Math.max(y - prevY, 0);
-    hv += x * dy;
-    prevY = Math.max(prevY, y);
-  }
-  return hv;
-}
-
-function weightedPick<T>(items: readonly T[], weights: readonly number[]): T {
-  const sum = weights.reduce((a, b) => a + (Number.isFinite(b) ? b : 0), 0);
-  if (sum <= 0) return items[Math.floor(Math.random() * items.length)]!;
-  let r = Math.random() * sum;
-  for (let i = 0; i < items.length; i++) {
-    const w = Number.isFinite(weights[i]!) ? weights[i]! : 0;
-    if (r < w) return items[i]!;
-    r -= w;
-  }
-  return items[items.length - 1]!;
-}
-
-function dominatesVectorEps(
-  a: Readonly<Record<string, number>>,
-  b: Readonly<Record<string, number>>,
-  eps = 0
-): boolean {
-  const keys = new Set([...Object.keys(a), ...Object.keys(b)]);
-  let atLeastAsGood = true;
-  let strictlyBetter = false;
-  for (const k of keys) {
-    const va = a[k] ?? 0;
-    const vb = b[k] ?? 0;
-    if (va + eps < vb) {
-      atLeastAsGood = false;
-      break;
-    }
-    if (va > vb + eps) strictlyBetter = true;
-  }
-  return atLeastAsGood && strictlyBetter;
-}
-
-function average(a: readonly number[]): number {
-  if (a.length === 0) return 0;
-  let s = 0;
-  for (const v of a) s += v;
-  return s / a.length;
-}
-
-function randomSubset<T>(arr: readonly T[], k: number): T[] {
-  if (k >= arr.length) return [...arr];
-  const picked = new Set<number>();
-  while (picked.size < k) picked.add(Math.floor(Math.random() * arr.length));
-  return Array.from(picked).map((i) => arr[i]!);
-}
-
-function selectCandidatePareto(S: number[][]): { index: number } {
-  // Build per-instance bests
-  const nCand = S.length;
-  const nInst = S[0]?.length ?? 0;
-  if (nCand <= 1 || nInst === 0) return { index: 0 };
-
-  // Best score per instance
-  const bestPerInst: number[] = new Array(nInst).fill(-Infinity);
-  for (let i = 0; i < nInst; i++) {
-    for (let k = 0; k < nCand; k++)
-      bestPerInst[i] = Math.max(bestPerInst[i], S[k]![i]!);
-  }
-
-  // Candidates that achieve best on at least one instance
-  const appears: number[] = new Array(nCand).fill(0);
-  for (let i = 0; i < nInst; i++) {
-    for (let k = 0; k < nCand; k++)
-      if (S[k]![i]! === bestPerInst[i]) appears[k]! += 1;
-  }
-
-  // Remove dominated candidates: if A is <= B for all i and < for at least one
-  const dominated = new Array(nCand).fill(false);
-  for (let a = 0; a < nCand; a++) {
-    for (let b = 0; b < nCand; b++) {
-      if (a === b) continue;
-      let allLe = true;
-      let strictlyLt = false;
-      for (let i = 0; i < nInst; i++) {
-        if (S[a]![i]! > S[b]![i]!) allLe = false;
-        if (S[b]![i]! > S[a]![i]!) strictlyLt = true;
-        if (!allLe) break;
-      }
-      if (allLe && strictlyLt) {
-        dominated[a] = true;
-        break;
-      }
-    }
-  }
-
-  // Build sampling weights from non-dominated set
-  const weights: number[] = [];
-  const indices: number[] = [];
-  for (let k = 0; k < nCand; k++) {
-    if (!dominated[k] && appears[k] > 0) {
-      indices.push(k);
-      weights.push(appears[k]);
-    }
-  }
-  if (indices.length === 0) return { index: Math.floor(Math.random() * nCand) };
-
-  const sumW = weights.reduce((a, b) => a + b, 0);
-  let r = Math.random() * sumW;
-  for (let j = 0; j < indices.length; j++) {
-    if (r < weights[j]!) return { index: indices[j]! };
-    r -= weights[j]!;
-  }
-  return { index: indices[indices.length - 1]! };
 }
